Rename `mock` to `stub`
