+++ src/lib/stub_api.js
