+++ test/stub_api.js
@@ -1,10 +1,10 @@
 import request from 'supertest';
 import app from '../src/app';
 
-describe('GET /mock/api/session/', () => {
+describe('GET /stub/api/session/', () => {
     it('respond with json', (done) => {
         request(app)
-        .get('/mock/api/session')
+        .get('/stub/api/session')
         .expect('Content-Type', /json/)
         .expect(200, done);
     });
